+++ src/app/locales/en/components/-browser-columns-name/filesystem.js
@@ -20,9 +20,11 @@ export default {
     ctime: 'Last time the item\'s metadata was changed, e.g. permissions. It\' s also always updated when the modification time changes.',
     creationTime: 'Time the item was created.',
     posixPermissions: 'Permissions regulating data access rights. It can be expressed either by the standard POSIX permissions system or by the Access Control List (ACL). If there is an ACL specified, the POSIX permissions are ignored.<br>Click on the cell to see more details.',
+    xattr: 'The extended attribute value for key: <strong>{{key}}</strong>',
   },
   subname: {
     modification: 'content',
     ctime: 'metadata',
+    xattr: 'xattr',
   },
 };