+++ src/app/locales/en/components/columns-configuration-popover.js
@@ -2,4 +2,15 @@ import columnItem from './columns-configuration-popover/column-item';
 
 export default {
   columnItem,
+  addNewColumn: 'New xattr column...',
+  addNewColumnHeader: 'New xattr column',
+  xattrNameLabel: 'Extended attribute key:',
+  columnNameLabel: 'Column label:',
+  addBtn: 'Create',
+  cancelBtn: 'Cancel',
+  modifyColumn: 'Modify xattr column',
+  modifyBtn: 'Apply',
+  backTooltip: 'Back',
+  dropdownPlaceholder: 'Enter an xattr key...',
+  customKeyPlaceholder: 'Custom xattr key...',
 };