+++ src/app/locales/en/components/modals.js
@@ -3,6 +3,7 @@ import externalSymlinkModal from './modals/external-symlink-modal';
 import recordListSelectorModal from './modals/record-list-selector-modal';
 import runWorkflowModal from './modals/run-workflow-modal';
 import shareModal from './modals/share-modal';
+import xattrModal from './modals/xattr-modal';
 
 export default {
   atmTaskExecutionPodsActivityModal,
@@ -10,4 +11,5 @@ export default {
   recordListSelectorModal,
   runWorkflowModal,
   shareModal,
+  xattrModal,
 };