+++ src/app/models/file.js
@@ -519,6 +519,12 @@ export default Model.extend(
     creationTime: attr('number'),
 
     /**
+     * The object contains the requested xattr attributes, with property names
+     * corresponding to xattr keys where periods have been replaced by dashes.
+     */
+    xattrs: attr('object', { defaultValue: () => {} }),
+
+    /**
      * One of: `posix`, `acl`. Cannot be modified
      */
     activePermissionsType: attr('string'),