+++ src/app/utils/filesystem-browser-model.js
@@ -362,6 +362,11 @@ export default BaseBrowserModel.extend(...mixins, {
     },
   ),
 
+  /**
+   * @type {boolean}
+   */
+  hasXattrColumnsSupport: true,
+
   parentDirRequirement: computed(
     'dir',
     'browserFilesProperties',
@@ -380,7 +385,7 @@ export default BaseBrowserModel.extend(...mixins, {
   listingRequirement: computed(
     'dir',
     'listedFilesProperties',
-    function parentDirRequirement() {
+    function listingRequirement() {
       if (!this.dir) {
         return;
       }
@@ -442,7 +447,7 @@ export default BaseBrowserModel.extend(...mixins, {
     'media.isMobile',
     'browserFilesProperties',
     `columnsConfiguration.columns.{${columnsRequirementsDependencies}}`,
-    'columnsConfiguration.isMounted',
+    'columnsConfiguration.listedFilesProperties',
     function listedFilesProperties() {
       const listedFilesPropertySet = new Set([
         ...this.browserFilesProperties,
@@ -464,38 +469,9 @@ export default BaseBrowserModel.extend(...mixins, {
           // TODO: VFS-11449 optional file size fetch
           .add('mtime');
       } else {
-        const columnRequirementsEnableProperty = this.columnsConfiguration.isMounted ?
-          'isVisible' : 'isEnabled';
-        const columns = this.columnsConfiguration.columns;
-        if (columns.size?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('size');
-        }
-        if (columns.modification?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('mtime');
-        }
-        if (columns.owner?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('owner');
-        }
-        if (columns.replication?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('localReplicationRate');
-        }
-        if (columns.qos?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('aggregateQosStatus');
-        }
-        if (columns.posixPermissions?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('posixPermissions');
-        }
-        if (columns.atime?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('atime');
-        }
-        if (columns.ctime?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('ctime');
-        }
-        if (columns.creationTime?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('creationTime');
-        }
-        if (columns.fileId?.[columnRequirementsEnableProperty]) {
-          listedFilesPropertySet.add('fileId');
+        const filesProperties = this.columnsConfiguration.listedFilesProperties;
+        for (const property of filesProperties) {
+          listedFilesPropertySet.add(property);
         }
       }
       return [...listedFilesPropertySet.values()];
@@ -1386,6 +1362,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: 140,
             hasSubname: false,
             hasTooltip: false,
+            type: 'basic',
+            fileProperty: 'size',
           });
           break;
         case 'modification':
@@ -1395,6 +1373,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: columnsTimesWidth,
             hasSubname: true,
             hasTooltip: true,
+            type: 'basic',
+            fileProperty: 'mtime',
           });
           break;
         case 'owner':
@@ -1404,6 +1384,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: 200,
             hasSubname: false,
             hasTooltip: false,
+            type: 'basic',
+            fileProperty: 'owner',
           });
           break;
         case 'replication':
@@ -1413,6 +1395,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: 160,
             hasSubname: false,
             hasTooltip: true,
+            type: 'basic',
+            fileProperty: 'localReplicationRate',
           });
           break;
         case 'qos':
@@ -1422,6 +1406,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: 100,
             hasSubname: false,
             hasTooltip: true,
+            type: 'basic',
+            fileProperty: 'aggregateQosStatus',
           });
           break;
         case 'atime':
@@ -1431,6 +1417,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: columnsTimesWidth,
             hasSubname: false,
             hasTooltip: true,
+            type: 'basic',
+            fileProperty: 'atime',
           });
           break;
         case 'ctime':
@@ -1440,6 +1428,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: columnsTimesWidth,
             hasSubname: true,
             hasTooltip: true,
+            type: 'basic',
+            fileProperty: 'ctime',
           });
           break;
         case 'creationTime':
@@ -1449,6 +1439,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: columnsTimesWidth,
             hasSubname: false,
             hasTooltip: true,
+            type: 'basic',
+            fileProperty: 'creationTime',
           });
           break;
         case 'fileId':
@@ -1458,6 +1450,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: 120,
             hasSubname: false,
             hasTooltip: false,
+            type: 'basic',
+            fileProperty: 'fileId',
           });
           break;
         case 'posixPermissions':
@@ -1467,6 +1461,8 @@ export default BaseBrowserModel.extend(...mixins, {
             width: 150,
             hasSubname: false,
             hasTooltip: true,
+            type: 'basic',
+            fileProperty: 'posixPermissions',
           });
           break;
         default:
@@ -1475,6 +1471,7 @@ export default BaseBrowserModel.extend(...mixins, {
     }
     return ColumnsConfiguration.create({
       configurationType: this.browserPersistedConfigurationKey,
+      hasXattrSettings: this.hasXattrColumnsSupport,
       columns,
       columnsOrder,
       firstColumnWidth: 380,